move parity tests to test_topoplot.py
